Refactor item tooltip to shared module
