+++ src/features/loot/ui/lootTab.js
@@ -1,6 +1,7 @@
 import { S } from '../../../shared/state.js';
 import { getSessionLoot } from '../selectors.js';
 import { forfeitSessionLoot } from '../mutators.js';
+import { showItemDetails as showDetails } from '../../../shared/utils/itemTooltip.js';
 
 export function updateLootTab(state = S) {
   const list = document.getElementById('sessionLootList');
@@ -10,6 +11,7 @@ export function updateLootTab(state = S) {
     const row = document.createElement('div');
     row.className = 'loot-row';
     row.textContent = `${item.qty || 1} ${item.key}`;
+    row.addEventListener('click', (e) => showDetails(item, e));
     list.appendChild(row);
   });
 }
